fix: mentor's remarks
