+++ pointer-poker/src/App.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { CSSTransition, TransitionGroup } from 'react-transition-group';
-import { BrowserRouter as Router, Redirect, Route, Switch, useLocation } from 'react-router-dom';
+import { BrowserRouter as Router} from 'react-router-dom';
 // import ComponentForCheckingBackend from './pages/CheckngBackEnd/ComponentForCheckingBackend';
-import Main from './pages/Main/Main';
-
 import './styles.scss';
+import PagesAll from './pages/pagesAll/PagesAll';
 
 export const App = () => {
   return (
@@ -20,32 +18,3 @@ export const App = () => {
 
 export default App;
 
-const PagesAll = () => {
-  const location = useLocation();
-
-  return (
-    <div className="pages">
-      <TransitionGroup>
-        <CSSTransition timeout={300} classNames="page" key={location.key}>
-          <Switch location={location}>
-            <Route exact path="/">
-              <Main />
-            </Route>
-            {/* <Route exact path="/about">
-              <About />
-            </Route>
-            <Route path="/details/:title">
-              <Details />
-            </Route>
-            <Route path="/error">
-              <NotFound />
-            </Route>
-            <Redirect to="/error">
-              <NotFound />
-            </Redirect> */}
-          </Switch>
-        </CSSTransition>
-      </TransitionGroup>
-    </div>
-  );
-};
