+++ pointer-poker/src/pages/Main/Main.scss
@@ -1,8 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
 .wrapper-main_page {
   display: flex;
   flex-direction: column;
