+++ pointer-poker/src/pages/contentMainPage/contentMainPage.scss
