+++ pointer-poker/src/pages/pagesAll/PagesAll.tsx
@@ -1,0 +1,23 @@
+import { Route, Switch, useLocation } from "react-router-dom";
+import { CSSTransition, TransitionGroup } from "react-transition-group";
+import Main from "../Main/Main";
+
+const PagesAll = () => {
+  const location = useLocation();
+
+  return (
+    <div className="pages">
+      <TransitionGroup>
+        <CSSTransition timeout={300} classNames="page" key={location.key}>
+          <Switch location={location}>
+            <Route exact path="/">
+              <Main />
+            </Route>
+          </Switch>
+        </CSSTransition>
+      </TransitionGroup>
+    </div>
+  );
+};
+
+export default PagesAll;
